feature: rename uefi/hooks to uefi/services
